+++ imdb/schema.js
@@ -9,20 +9,22 @@ import MovieType from './object_types.js';
 
 // variables
 const baseUrl = 'http://www.omdbapi.com/';
+const byTitleParam = '?t='
+const apikey = '&apikey=';
 
 module.exports = new GraphQLSchema({
   query: new GraphQLObjectType({
-    name: 'Query',
-    description: 'GraphQLObjectType Query to fetch movie by imdb ID',
+    name: 'Search',
+    description: 'GraphQLObjectType: Search for a specific movie title on Omdb API',
 
     fields: () => ({
       movie: {
         type: MovieType,
         args: {
-          id: { type: GraphQLString }
+          title: { type: GraphQLString }
         },
         resolve: (root, args) => fetch(
-          `${baseUrl}?i=${args.id}&apikey=${global.gConfig.omdb_api_key}`
+          baseUrl + byTitleParam + args.title + apikey + global.gConfig.omdb_api_key
         ).then(response => response.json())
       }
     })